requirements.txt
# Azure Functions SDK
azure-functions

# Add your dependencies below this line
# Example:
# requests
# pandas
